+++ spec/Users/UserSpec.php
@@ -12,7 +12,7 @@ class UserSpec extends ObjectBehavior
 {
     public function let(AlmaClient $almaClient)
     {
-        $this->beConstructedWith($almaClient, '123456');
+        $this->beConstructedWith($almaClient, '12345');
 
         $almaClient->getJSON(Argument::containingString('12345'), Argument::any())
             ->willReturn(SpecHelper::getDummyData('user_response.json'));
@@ -23,10 +23,33 @@ class UserSpec extends ObjectBehavior
         $this->shouldHaveType(User::class);
     }
 
+    public function it_has_primary_id()
+    {
+        $this->primaryId->shouldBe('12345');
+    }
+
+    public function it_has_barcode()
+    {
+        $this->barcode->shouldBe('ub54321');
+    }
+
+    public function it_has_barcodes()
+    {
+        $this->barcodes->shouldBe(['ub54321', 'ntb12897787']);
+    }
+
+    public function it_has_university_id()
+    {
+        $this->universityId->shouldBe('test@uio.no');
+    }
+
+    public function it_has_university_ids()
+    {
+        $this->universityIds->shouldBe(['test@uio.no']);
+    }
+
     public function it_has_identifiers()
     {
-        $this->primary_id->shouldBe('12345');
-        $this->getUniversityId()->shouldBe('test@uio.no');
-        $this->getBarcode()->shouldBe(null);
+        $this->identifiers->all()->shouldBe(['12345', 'ub54321', 'ntb12897787', 'test@uio.no']);
     }
 }